Refactored compiler.
The previous code had a lot of boilerplate copied around to preserve a
data-structure. The new code uses a couple of helper methods to
manipulate this structure in a few helpful ways.

Further refactoring to turn the data structure into an object with
methods would make the code even better, shifting the generator to a
jQuery-like fluent syntax.