NEW: Add basic variable declaration support.
Variables can now be declared and used in expressions; the main purpose
of the variable declaration is to infer its type.

In the future, I’d like to add a AST processor that will help with type
inferencing. But for now, explicit declaration will work.

I’ve also made argument orders in the AST construction a bit more
consistent: it’s value/name, then type.