NEW: Improve LLVM type handling be strict about string length.
LLVM has two types for a string, i8* and [12 x i8], which is used to
indicate a 12-byte string, typically when provided as a constant.

In most cases, the constant string will need to be converted to a
pointer using getelementptr.

This change ensures that this happens consistently. It occurs a few
more times than in the previous code, but this shouldn’t have too
adverse an impact.